requirements.txt
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
numpy==1.21.0
pillow==8.3.1
